+++ rdm_binderhub_jlabextension/__init__.py
@@ -1,25 +1,16 @@
-
-import json
-from pathlib import Path
-
 from tornado.ioloop import IOLoop
 
-from .config import save_env_to_home
-from ._version import __version__
-
-HERE = Path(__file__).parent.resolve()
-
-with (HERE / "labextension" / "package.json").open() as fid:
-    data = json.load(fid)
-
-def _jupyter_labextension_paths():
-    return [{
-        "src": "labextension",
-        "dest": data["name"]
-    }]
-
-
-
+try:
+    from ._version import __version__
+except ImportError:
+    # Fallback when using the package in dev mode without installing
+    # in editable mode with pip. It is highly recommended to install
+    # the package from a stable release or in editable mode: https://pip.pypa.io/en/stable/topics/local-project-installs/#editable-installs
+    import warnings
+    warnings.warn("Importing 'rdm_binderhub_jlabextension' outside a proper installation.")
+    __version__ = "dev"
+from .handlers import setup_handlers
+from .config import SyncToRDM, save_env_to_home
 from . import server
 
 
@@ -33,6 +24,13 @@ def _jupyter_nbextension_paths():
                  src="nbextension",
                  dest="rdm_binderhub_jlabextension",
                  require="rdm_binderhub_jlabextension/main")]
+
+
+def _jupyter_labextension_paths():
+    return [{
+        "src": "labextension",
+        "dest": "rdm-binderhub-jlabextension"
+    }]
 
 
 def _jupyter_server_extension_points():
@@ -49,10 +47,8 @@ def _load_jupyter_server_extension(server_app):
     server_app: jupyterlab.labapp.LabApp
         JupyterLab application instance
     """
-    server.register_routes(server_app, server_app.web_app)
+    setup_handlers(SyncToRDM(parent=server_app), server_app.web_app)
     IOLoop.current().add_callback(server.task_worker.watch_queue)
     save_env_to_home()
-    server_app.log.info("Registered GakuNin RDM extension at URL path /rdm-binderhub")
-
-# For backward compatibility with notebook server - useful for Binder/JupyterHub
-load_jupyter_server_extension = _load_jupyter_server_extension
+    name = "rdm_binderhub_jlabextension"
+    server_app.log.info(f"Registered {name} server extension")