+++ rdm_binderhub_jlabextension/handlers.py
@@ -1,0 +1,13 @@
+from jupyter_server.utils import url_path_join
+
+from .server import FilesHandler
+
+
+def setup_handlers(app, web_app):
+    host_pattern = ".*$"
+
+    base_url = web_app.settings["base_url"]
+    handler_options = app.get_config()
+    route_pattern = url_path_join(base_url, "rdm-binderhub-jlabextension", "files")
+    handlers = [(route_pattern, FilesHandler, handler_options)]
+    web_app.add_handlers(host_pattern, handlers)