+++ src/handler.ts
@@ -17,7 +17,7 @@ export async function requestAPI<T>(
   const settings = ServerConnection.makeSettings();
   const requestUrl = URLExt.join(
     settings.baseUrl,
-    'rdm-binderhub', // API Namespace
+    'rdm-binderhub-jlabextension', // API Namespace
     endPoint
   );
 
@@ -25,14 +25,22 @@ export async function requestAPI<T>(
   try {
     response = await ServerConnection.makeRequest(requestUrl, init, settings);
   } catch (error) {
-    throw new ServerConnection.NetworkError(error);
+    throw new ServerConnection.NetworkError(error as any);
   }
 
-  const data: any = await response.text();
+  let data: any = await response.text();
+
+  if (data.length > 0) {
+    try {
+      data = JSON.parse(data);
+    } catch (error) {
+      console.log('Not a JSON response body.', response);
+    }
+  }
 
   if (!response.ok) {
     throw new ServerConnection.ResponseError(response, data.message || data);
   }
 
-  return JSON.parse(data);
+  return data;
 }