+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 <!-- <START NEW CHANGELOG ENTRY> -->
 
+- Made Migration to JupyterLab 4.x
+
 <!-- <END NEW CHANGELOG ENTRY> -->